Create Final.sql
In PGAdmin, create a new database, and create a new schema called "final_project". Then, right-click on the database (in the object explorer), and click restore. From there, you will select this file. The skeleton of this database should now be in the final_project schema--tables, columns, and primary keys.